New multiply add module for PE implementing convergent round to even
Convergent round to even was extended in a copy of the `alpaca_multadd` module.
This implements convergent round to even using the `lsb correction technique`
and programming for dsp recommendations as described in Xilinx UG193.

In this methodology the idea is to use the DSP48Es C, Carryin and Pattern Detect
capabilities to implement the round. Then on the next clock cycle the FPGA
fabric flips the LSB if the pattern detect output is high.

The convergent round block uses the output `fp_data` interface to determine the
output width and scaling. Right now we scale by shifting right then keep the
bottom `width` bits corresponding to the width of the output. In the future it
may be a good idea to extend this block to have a variable scale input. Or to
create a seperate convergent round block that can accept an arbitrary scaling.