Extend convergent round to the cx multiply for output of parallel fft
Exact same idea as last commit for applying convergent rounding to the ouput of
the PEs but now to the output of the twiddle factor recombination step of the
parallel fft.

There is something strange going on when the FFT scaling schedule is not the
default scaling schedule. For this case with M=128 the parallel ffts with N=64
that scaling schedule is [0b10, 0b10, 0b10, 0b10]. If I do something other than
this than there is an artifact in the output on the opposite side of the
spectrum that was similar to when I first doing the twiddle factor recombination
and not aligning the implicit binary point in the addition step.

Now that the output is at 16 bits instead of the 40 bits and every multiply add
is now pipelined this last step should be the last capability needed and we can
now extend to ALPACA specifications and pending those results begin to build to
hardware and see what optimizations need to be made from there.