Full single antenna alpaca spec ospfb runs through synthesis with one timing issue
Took the full alpaca spec that worked in simulation and ran it in non-project
mode through synthesis, optimzation, implementation and place and route. I am
mostly satisfied with the results.

Changed the way the twiddle factors are passed to match that of the FIR taps by
passing as a parameter from a pacakge declaration. Then they are set to a typed
version of the ram. This works in vivado synthesis without needing to use the
verbose declaration by using the $bits system method to get the correct type
that is required when loading a memory file with $readmemh and the associated
"malformed readmem call" error that vivado synthesis throws.

Also added widths to the interface types for module hierarchy to use and create
subsequent interfaces and perform arithmetic. The error was that Vivado
synthesis requires that parameter declarations be const and so doing
```
localparam width = $bits(myint.data_t)
```
to use to set widths was failing.

I was really pleased to see the synthesis report show most of my hardware in
schematic view and that it correctly inferred all the correct elements. Even
without too many optimzation warnings. There was only one and that was when
loading a twiddle factor out of ram into the cx mult add module. It was easy to
extend this by just adding one to the overall module latency and cross boundary
optimzation picked that extra register up as the output register of the twiddle
ROM.

While the output register does not show up on the schematic view next to the ROM
the BRAM parameters in the schematic view shows that the DOA register is enabled
and according to Xilinx forum posts this indicates that it is indeed present in
the design.

The BRAM parameters for both the FIR taps and twiddle ROMs show initialization
values and so I am pretty confident that those have been passed in correctly to
the modules and will be present at startup correctly.

Currently there is timing errors at the crossing between clock domains from the
ADC to the OSPFB. I am partially thinking this is mostly due to the fact that I
don't actually have the correct clocking infrastrucutre in the desisgn or any
input output location constraints and so when the place and route creates the
clock eveyrthing is spread out rather far and so I am planning on revisiting the
clock timing errors later. Because since everything passes timing in the OSPFB
and this clock error is within the Xilinx FIFO (nothing I wrote) then we should
be OK.