Reset synchronous to dsp clk and processing
I have been worried about not handling the reset across clock domains. But,
unless I am wrong, handling resets in the ospfb may already be done sufficiently
handled for now until more functionality of the core is needed.

To do this though requires I specify the behavior. The OSPFB consits of a
Xilinx XPM dual-clock fifo followed by my custom processing core. The XPM fifo
accepts data from the adc at one clock and processes on the output at the other.
The XPM fifo accepts an asynchronous reset on the slave domain (the adc, slower
clock) and the polyphase fir side fo the core accepts a synchronous input. If I
reset my design synchronous to the dsp processing this reset will be asynch.
with resepct to the xpm fifo. But, since that fifo is designed for that it
shoudl handle reset and come out of it just fine. Thus, my entire design will
come out of reset fine.

So it seems to make the most sense to make that design specification and deal
with more complexity down the road as it is needed.